+++ rust/bridge/shared/types/src/net/remote_config.rs
@@ -105,8 +105,6 @@ pub enum RemoteConfigKey {
     /// How long to wait for a response to a chat request before checking whether the connection is
     /// still active.
     ChatRequestConnectionCheckTimeoutMilliseconds => "chatRequestConnectionCheckTimeoutMillis",
-    /// Whether to disable the Nagle algorithm (sets TCP_NODELAY).
-    DisableNagleAlgorithm => "disableNagleAlgorithm",
     /// If set, unauth chat connections (only!) will connect over H2.
     UseH2ForUnauthChat => "useH2ForUnauthChat",