+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.88.2
 
+- The `disableNagleAlgorithm` remote config flag has been removed, as the experiment has been deployed successfully.